requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
psycopg2-binary>=2.9.9
sqlparse>=0.5.0
requests>=2.32.0
